backend/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
python-dotenv==1.0.0
faiss-cpu==1.7.4
pypdf==3.17.4
docx2txt==0.8
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6